LayoutTests/fast/forms/form-disassociation-attributes-expected.txt

This test checks the form attribute of the form-associated elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



- Check if a form and a control are disassociated when they are removed from the document together.
PASS owner.elements.length is 1
PASS owner.elements.length is 0
PASS successfullyParsed is true

TEST COMPLETE
